src/requirements.txt
ansible==6.5.0
ansible-core==2.13.5
ansible-runner==2.2.1
cffi==1.15.1
cryptography==38.0.1
deepdiff==6.2.1
docutils==0.19
Jinja2==3.1.2
lockfile==0.12.2
MarkupSafe==2.1.1
ordered-set==4.1.0
packaging==21.3
pexpect==4.8.0
ptyprocess==0.7.0
pycparser==2.21
pyparsing==3.0.9
python-daemon==2.3.2
PyYAML==6.0
resolvelib==0.8.1
six==1.16.0
watchdog==2.1.9